Collapsable deploy panel
