+++ web/dashboard/src/redux/eventHandlers/eventHandlers.js
@@ -92,6 +92,16 @@ export function envStackUpdated(state, envName, payload) {
   return state;
 }
 
+export function openDeployPanel(state) {
+  state.deployPanelOpen = true
+  return state;
+}
+
+export function closeDeployPanel(state) {
+  state.deployPanelOpen = false
+  return state;
+}
+
 export function envPullRequests(state, payload) {
   for (const [envName, pullRequests] of Object.entries(payload)) {
     if (!state.envs.some(env => env.name === envName)) {
