+++ web/dashboard/src/redux/redux.js
@@ -41,6 +41,9 @@ export const ACTION_TYPE_ALERTS = 'alerts'
 export const ACTION_TYPE_POPUPWINDOWSUCCESS = 'popupWindowSaved';
 export const ACTION_TYPE_POPUPWINDOWRESET = 'popupWindowReset';
 
+export const ACTION_TYPE_OPEN_DEPLOY_PANEL = 'openDeployPanel';
+export const ACTION_TYPE_CLOSE_DEPLOY_PANEL = 'closeDeployPanel';
+
 export const EVENT_AGENT_CONNECTED = 'agentConnected';
 export const EVENT_AGENT_DISCONNECTED = 'agentDisconnected';
 export const EVENT_ENVS_UPDATED = 'envsUpdated';
@@ -96,7 +99,8 @@ export const initialState = {
     errorList: null
   },
   podLogs: {},
-  users: []
+  users: [],
+  deployPanelOpen: false
 };
 
 export function rootReducer(state = initialState, action) {
@@ -171,6 +175,10 @@ export function rootReducer(state = initialState, action) {
        return podEventHandlers.clearPodLogs(state, action.payload)
     case ACTION_TYPE_ENVUPDATED:
       return eventHandlers.envStackUpdated(state, action.name, action.payload)
+    case ACTION_TYPE_OPEN_DEPLOY_PANEL:
+      return eventHandlers.openDeployPanel(state)
+    case ACTION_TYPE_CLOSE_DEPLOY_PANEL:
+      return eventHandlers.closeDeployPanel(state)
     default:
       console.log('Could not process redux event: ' + JSON.stringify(action));
       return state;
